+++ NEWS.md
@@ -1,3 +1,6 @@
 # excelsior (development version)
 
-* Initial CRAN submission.
+- `row_finder()`: finds row number based on an "anchor" cell of known column and pattern.
+- `as_numeric_smart()`: as.numeric(), but correctly handles cases of numbers with commas in them, numbers ending in %s.
+- `copy_columns()`: copies specified sets of columns from start_row to end_row from one file into another, with highlighting of cells copied and cells with changed values
+- addition of `validate_*()` for input validation